create utils file
