+++ libs/trainer/mlpe/trainer/trainer.py
@@ -8,29 +8,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 import torch
 
+from mlpe.trainer.utils import model_memory, number_trainable_parameters
+
 if TYPE_CHECKING:
     from nflows import transforms
-
-
-def step_scheduler(scheduler: torch.optim.lr_scheduler, metric: float = None):
-    if isinstance(scheduler, torch.optim.lr_scheduler.ReduceLROnPlateau):
-        scheduler.step(metric)
-    else:
-        scheduler.step()
-
-
-def number_trainable_parameters(model: torch.nn.Module):
-    """Return a list of trainable parameters in a model"""
-    num = sum(p.numel() for p in model.parameters() if p.requires_grad)
-    num += sum(p.numel() for p in model.buffers() if p.requires_grad)
-    return num
-
-
-def model_memory(model: torch.nn.Module):
-    """Return the memory used by a model in Mega bytes"""
-    memory = sum(p.numel() * p.element_size() for p in model.parameters())
-    memory += sum(p.numel() * p.element_size() for p in model.buffers())
-    return memory / 1e6
 
 
 def train_for_one_epoch(
